fix: improve performance by swapping pouch-queue for level-queue
Oops sorry future me for the massive commit!

Unfortunately I had to touch learning-strategy.js. This happened
because we can only have one instance of levelup at a time (pointing to
the same location in the file system). To solve this I added a method
to async-deck-loader that would return the instance of the DB we should
communicate with. Then all of the queues use this same DB.